--- a/packages/protons-runtime/package.json
+++ b/packages/protons-runtime/package.json
@@ -148,8 +148,8 @@
     "release": "aegir release"
   },
   "dependencies": {
-    "protobufjs": "^7.0.0",
-    "uint8arraylist": "^2.3.2"
+    "uint8arraylist": "^2.3.2",
+    "uint8arrays": "^3.1.0"
   },
   "devDependencies": {
     "aegir": "^37.0.5"