--- a/packages/protons-runtime/src/encode.ts
+++ b/packages/protons-runtime/src/encode.ts
@@ -1,25 +1,9 @@
 import type { Codec } from './codec.js'
-import pb from 'protobufjs'
-
-const Writer = pb.Writer
-
-// monkey patch the writer to add native bigint support
-const methods = [
-  'uint64', 'int64', 'sint64', 'fixed64', 'sfixed64'
-]
-methods.forEach(method => {
-  // @ts-expect-error
-  const original = Writer.prototype[method]
-  // @ts-expect-error
-  Writer.prototype[method] = function (val: bigint): pb.Writer {
-    return original.call(this, val.toString())
-  }
-})
+import { createWriter } from './utils/writer.js'
 
 export function encodeMessage <T> (message: T, codec: Codec<T>): Uint8Array {
-  const w = Writer.create()
+  const w = createWriter()
 
-  // @ts-expect-error
   codec.encode(message, w, {
     lengthDelimited: false
   })