… packages/protons-runtime/src/utils.ts
… packages/protons-runtime/src/utils/writer.ts