Ports just the parts of protobufjs we use to typescript to avoid the extra deps and integrates the changes from https://github.com/protobufjs/protobuf.js/pull/1557 to have native BigInt support.