@achingbrain 2022-08-10
This cannot be merged until https://github.com/nodejs/node/issues/44186 as resolved as it introduces a performance regression due to the use of small, short-lived ESM classes.