@wemeetagain 2023-09-08
@achingbrain can this be revisited?
It seems that the problem was more so related to the style of import (eg: named vs default vs namespace) rather than esm vs cjs.