@wemeetagain 2023-09-11
This will open the door for more performance improvements (eg: using a more optimized varint reader / writer)

This will also make adding additional types more straightforward (eg: 64 bit number types associated with https://github.com/ipfs/protons/issues/112)