@wemeetagain 2023-09-13
This appears to have no performance penalty vs the current master.
current master
```
pbjs x 8,051 ops/sec ±0.60% (91 runs sampled)
protons x 7,993 ops/sec ±0.45% (91 runs sampled)
protobuf.js x 8,182 ops/sec ±0.71% (91 runs sampled)
```

this PR
```
pbjs x 8,081 ops/sec ±0.69% (89 runs sampled)
protons x 8,353 ops/sec ±0.28% (95 runs sampled)
protobufjs x 8,376 ops/sec ±0.29% (95 runs sampled)
```